+++ notifications-20241225/ChangeLog.md
@@ -1,3 +1,14 @@
+2025-09-18 Version: 1.0.2
+- Update API DeleteAllMessage: add request parameters GroupCode.
+- Update API ReadAllMessage: add request parameters GroupCode.
+- Update API ReadMessageContent: add request parameters GroupCode.
+- Update API ReadMessageContent: add request parameters History.
+- Update API ReadMessageList: add request parameters GroupCode.
+- Update API ReadMessageList: add request parameters History.
+- Update API ReadMessageList: add response parameters Body.Data.Rows.$.CategoryCode.
+- Update API ReadNumGroupTotal: add response parameters Body.Data.$.GroupCode.
+
+
 2025-08-07 Version: 1.0.1
 - Update API ReadMessageList: add response parameters Body.Data.Rows.$.Titleh.
 - Update API ReadNumGroupTotal: add request parameters Title.